+++ apps/explorer/lib/explorer/application.ex
@@ -131,7 +131,8 @@ defmodule Explorer.Application do
         configure(Explorer.Migrator.AddressTokenBalanceTokenType),
         configure(Explorer.Migrator.SanitizeMissingBlockRanges),
         configure(Explorer.Migrator.SanitizeIncorrectNFTTokenTransfers),
-        configure(Explorer.Migrator.TokenTransferTokenType)
+        configure(Explorer.Migrator.TokenTransferTokenType),
+        configure_chain_type_dependent_process(Explorer.Chain.Cache.ValidatorStabilityCounter, "stability")
       ]
       |> List.flatten()
 
@@ -148,7 +149,8 @@ defmodule Explorer.Application do
         Explorer.Repo.Shibarium,
         Explorer.Repo.Suave,
         Explorer.Repo.BridgedTokens,
-        Explorer.Repo.Filecoin
+        Explorer.Repo.Filecoin,
+        Explorer.Repo.Stability
       ]
     else
       []
@@ -169,6 +171,14 @@ defmodule Explorer.Application do
 
   defp configure(process) do
     if should_start?(process) do
+      process
+    else
+      []
+    end
+  end
+
+  defp configure_chain_type_dependent_process(process, chain_type) do
+    if Application.get_env(:explorer, :chain_type) == chain_type do
       process
     else
       []